Extract XiClient implementation from AppDelegate
